Fix flicker upon book load
The flicker is caused by the bottom nav bar appearing in the middle of
the screen.

We need to do 2 things:
1. Justify the navbar to the bottom.
2. Add an animation to the navigation buttons to slide up.

The second is necessary because for content that is longer than the
page, we still see the navbar flicker from the bottom to off-screen.
The translating upwards effectively hides that problem.