Fix bug in local updates
Previously reading from serverTimestamp field, which is wrong
because we might read from remote, and try to apply a change, and there
would be a serverTimestamp, even though it's a local write that needs
to be noted down and pushed.

To fix this, the storage adapter has to pass in a flag (symbol) to each
of the remote records to identify them as remote.